requirements.txt
# Django
Django==6.0.1
django-admin-interface==0.32.0
django-tinymce==5.0.0
django-advance-thumbnail==1.1.2
django-health-check==3.20.8
django-cleanup==9.0.0
django-select2==8.4.8
django-vite==3.1.0
django-bootstrap5==26.1
django-cache-cleaner==0.2.0
django-modeltrans==0.9.0
django-modeltrans-tabs==0.1.1
django-recaptcha==4.1.0
django-solo==2.5.1
django-admin-sortable2==2.3.1

# Environment variables
django-environ==0.12.0
python-dotenv==1.2.1

# prod server
gunicorn==23.0.0

# image manipulation
pillow==12.1.0

# Dataabse (psycopg[binary] cannot be locked to specific version number)
psycopg[binary]
